requirements_file.txt
# Core Streamlit and UI
streamlit==1.39.0
streamlit-chat==0.1.1

# Strands Agents SDK
strands-agents==1.0.0
strands-tools

# AWS SDK and Bedrock
boto3==1.35.36
botocore==1.35.36
aws-bedrock-agent-runtime-client
amazon-bedrock-client

# Google Gemini API
google-generativeai==0.8.3

# Additional utilities
python-dotenv==1.0.1
pydantic==2.9.2
typing-extensions==4.12.2

# Data processing and utilities
pandas==2.2.3
numpy==1.26.4
requests==2.32.3

# Async support
asyncio-throttle==1.0.2
aiohttp==3.10.10

# JSON and data handling
orjson==3.10.7
jsonschema==4.23.0

# Logging and monitoring
structlog==24.4.0
rich==13.9.2